Edit upload form based on Brian feedback
The page is now designed to look like the Balsamiq markup. Notably, the placeholder text for the recipe details textares instructs the user to place each item on it's own line. The review button does not do anythong yet. I need to figure out how to open a new page with a formatted recipe. I also got rid of the formfields because Brain was having trouble with them. currently, the different sections of the form are divs with labels. I don't know if this is easier with a screen reader yet, and I would love some input on that.